refactor: pull out MainDash into own file
